+++ src/main/java/com/limechain/network/protocol/grandpa/messages/consensus/GrandpaConsensusMessageReader.java
@@ -7,6 +7,7 @@ import io.emeraldpay.polkaj.scale.ScaleReader;
 import io.emeraldpay.polkaj.scale.reader.ListReader;
 import io.emeraldpay.polkaj.scale.reader.UInt64Reader;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class GrandpaConsensusMessageReader implements ScaleReader<GrandpaConsensusMessage> {
@@ -21,22 +22,24 @@ public class GrandpaConsensusMessageReader implements ScaleReader<GrandpaConsens
         switch (format) {
             case GRANDPA_SCHEDULED_CHANGE -> {
                 List<Authority> authorities = reader.read(new ListReader<>(new AuthorityReader()));
-                long delay = reader.readUint32();
+                BigInteger delay = BigInteger.valueOf(reader.readUint32());
 
                 grandpaConsensusMessage.setAuthorities(authorities);
                 grandpaConsensusMessage.setDelay(delay);
             }
             case GRANDPA_FORCED_CHANGE -> {
-                long additionalOffset = reader.readUint32();
+                BigInteger additionalOffset = BigInteger.valueOf(reader.readUint32());
                 List<Authority> authorities = reader.read(new ListReader<>(new AuthorityReader()));
-                long delay = reader.readUint32();
+                BigInteger delay = BigInteger.valueOf(reader.readUint32());
 
                 grandpaConsensusMessage.setAuthorities(authorities);
                 grandpaConsensusMessage.setDelay(delay);
                 grandpaConsensusMessage.setAdditionalOffset(additionalOffset);
             }
             case GRANDPA_ON_DISABLED -> grandpaConsensusMessage.setDisabledAuthority(new UInt64Reader().read(reader));
-            case GRANDPA_PAUSE, GRANDPA_RESUME -> grandpaConsensusMessage.setDelay(reader.readUint32());
+            case GRANDPA_PAUSE, GRANDPA_RESUME -> grandpaConsensusMessage.setDelay(
+                    BigInteger.valueOf(reader.readUint32())
+            );
         }
 
         return grandpaConsensusMessage;
